requirements.txt
# Core
pandas==2.1.4
python-dotenv==1.0.1
PyYAML==6.0.1
loguru==0.7.2

# Apache Spark
pyspark==3.4.1

# Kafka
kafka-python==2.0.2

# MongoDB
pymongo==4.6.1

# Dev Tools
jupyter==1.0.0
ipykernel==6.29.4